refactor(character): move dialog components to dialog subpackage
