+++ cmp-ttrpg-companion/composeApp/src/commonMain/kotlin/com/cyrillrx/rpg/character/presentation/component/dialog/CharacterEditDialogs.kt
@@ -1,4 +1,4 @@
-package com.cyrillrx.rpg.character.presentation.component
+package com.cyrillrx.rpg.character.presentation.component.dialog
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
